reorganize easy tests, remove depreciated auto_lr_find from baseline dag trainer
